fixed siging for standalone bets
Former-commit-id: e143dd11106a7d3e65b802ecfe4909c3233a3d96

+++ Scripts/build-electron-app.sh
@@ -1,40 +1,55 @@
 #!/bin/bash
 
-SRC="${SRCROOT}/../ElectronMainApp"
+ELECTRON="ElectronMainApp"
+SRC="${SRCROOT}/../${ELECTRON}"
 SHAREDSRC="${SRCROOT}/../Shared"
-
-mkdir -p $TARGET_TEMP_DIR/arm64
-mkdir -p $TARGET_TEMP_DIR/x86_64
-
-
-mkdir -vp "${SRC}/safari-ext/shared"
-cp -v "${BUILT_PRODUCTS_DIR}/libshared.a" "${SRC}/safari-ext/shared/" || exit 1
-rsync -avm --include='*.h' -f 'hide,! */' "${SHAREDSRC}/" "${SRC}/safari-ext/shared/"
-
-echo "Build electron app for arm64"
-bash ../Scripts/build-electron-app-one-arch.sh arm64 "$TARGET_TEMP_DIR/arm64" || exit 1
-
-echo "Build electron app for x86_64"
-bash ../Scripts/build-electron-app-one-arch.sh x86_64 "$TARGET_TEMP_DIR/x86_64" || exit 1
-
-echo "Create universal build"
-cd ../ElectronMainApp
-yarn make-universal-app "$TARGET_TEMP_DIR"
-
-echo "Copy converter binary"
-mkdir -p "${SRC}/../libs"
-cp "${SRC}/node_modules/safari-converter-lib/bin/ConverterTool" "${SRC}/../libs" || exit 1
-chmod +x "${SRC}/../libs/ConverterTool"
-
-echo "Processing ConverterTool"
-install_name_tool -add_rpath @executable_path/../Frameworks "${SRC}/../libs/ConverterTool" > /dev/null 2>&1 | echo -n
-install_name_tool -add_rpath @executable_path/../../Frameworks "${SRC}/../libs/ConverterTool" > /dev/null 2>&1 | echo -n
-
 APP_NAME="${PRODUCT_NAME}.app"
 APP="${TARGET_TEMP_DIR}/${APP_NAME}"
 
-echo "Hide executable file from xcodebuild"
-mv -f "$APP/Contents/MacOS/${PRODUCT_NAME}" "$APP/Contents/MacOS/${AG_HIDE_EXEC_PREFIX}${PRODUCT_NAME}" || exit 1
+
+TIME_MARKER="${TARGET_TEMP_DIR}/ElectronTimeMarker.touch"
+
+[ -f "$TIME_MARKER" ] && /usr/bin/find "${SRC}/" -newer "${TIME_MARKER}" | grep ${ELECTRON} > /dev/null
+
+if [[ $? == 0 ]] || ! [ -f "$TIME_MARKER" ]; then
+
+    echo "Rebuild electron started"
+
+    mkdir -p $TARGET_TEMP_DIR/arm64
+    mkdir -p $TARGET_TEMP_DIR/x86_64
+
+
+    mkdir -vp "${SRC}/safari-ext/shared"
+    cp -v "${BUILT_PRODUCTS_DIR}/libshared.a" "${SRC}/safari-ext/shared/" || exit 1
+    rsync -avm --include='*.h' -f 'hide,! */' "${SHAREDSRC}/" "${SRC}/safari-ext/shared/"
+
+    echo "Build electron app for arm64"
+    bash ../Scripts/build-electron-app-one-arch.sh arm64 "$TARGET_TEMP_DIR/arm64" || exit 1
+
+    echo "Build electron app for x86_64"
+    bash ../Scripts/build-electron-app-one-arch.sh x86_64 "$TARGET_TEMP_DIR/x86_64" || exit 1
+
+    echo "Create universal build"
+    cd ../ElectronMainApp
+    yarn make-universal-app "$TARGET_TEMP_DIR"
+
+    touch "${TIME_MARKER}"
+
+    echo "Copy converter binary"
+    mkdir -p "${SRC}/../libs"
+    cp "${SRC}/node_modules/safari-converter-lib/bin/ConverterTool" "${SRC}/../libs" || exit 1
+    chmod +x "${SRC}/../libs/ConverterTool"
+
+    echo "Processing ConverterTool"
+    install_name_tool -add_rpath @executable_path/../Frameworks "${SRC}/../libs/ConverterTool" > /dev/null 2>&1 | echo -n
+    install_name_tool -add_rpath @executable_path/../../Frameworks "${SRC}/../libs/ConverterTool" > /dev/null 2>&1 | echo -n
+
+    echo "Hide executable file from xcodebuild"
+    mv -f "$APP/Contents/MacOS/${PRODUCT_NAME}" "$APP/Contents/MacOS/${AG_HIDE_EXEC_PREFIX}${PRODUCT_NAME}" || exit 1
+
+fi
+
+echo "Copying electron app to: ${DST_DIR}"
 
 DST_DIR="${BUILT_PRODUCTS_DIR}"
 if [[ ${ACTION} == "install" ]]; then
@@ -43,7 +58,7 @@ if [[ ${ACTION} == "install" ]]; then
 fi
 
 rm -Rfv "${DST_DIR}/${APP_NAME}"
-cp -HRfp "${APP}" "${DST_DIR}"
+cp -HRfpv "${APP}" "${DST_DIR}"
 
 # Get rid of redundant asar files (need only one app.asar):
 # for the moment app.asar is path resolver for app-x64.asar and app-arm64.asar, that we don't need
